+++ source/extensions/omni.isaac.ros2_bridge/python/tests/test_rtx_sensor.py
@@ -1,11 +1,11 @@
-# Copyright (c) 2018-2021, NVIDIA CORPORATION.  All rights reserved.
-#
-# NVIDIA CORPORATION and its licensors retain all intellectual property
-# and proprietary rights in and to this software, related documentation
-# and any modifications thereto.  Any use, reproduction, disclosure or
-# distribution of this software and related documentation without an express
-# license agreement from NVIDIA CORPORATION is strictly prohibited.
-#
+__copyright__ = "Copyright (c) 2018-2022, NVIDIA CORPORATION. All rights reserved."
+__license__ = """
+NVIDIA CORPORATION and its licensors retain all intellectual property
+and proprietary rights in and to this software, related documentation
+and any modifications thereto. Any use, reproduction, disclosure or
+distribution of this software and related documentation without an express
+license agreement from NVIDIA CORPORATION is strictly prohibited.
+"""
 
 # NOTE:
 #   omni.kit.test - std python's unittest module with additional wrapping to add suport for async/await tests
@@ -13,19 +13,20 @@
 
 import omni.kit.test
 import omni.kit.commands
-import sys
+
 import carb.tokens
+import asyncio
+import math
 import numpy as np
 from pxr import Gf, UsdGeom, Usd, UsdPhysics, Sdf
 import omni.kit.commands
 import omni
 import omni.kit
-import omni.usd
 
 # Import extension python module we are testing with absolute import path, as if we are external user (other extension)
 from omni.isaac.sensor import _sensor
 from omni.syntheticdata import sensors
-import omni.kit.viewport.utility
+from omni.kit.viewport.utility import get_active_viewport
 from omni.isaac.core.utils.viewports import add_aov_to_viewport
 
 
@@ -44,7 +45,7 @@ def add_cube(stage, path, scale, offset, physics=False):
 
 
 # Having a test class dervived from omni.kit.test.AsyncTestCase declared on the root of module will make it auto-discoverable by omni.kit.test
-class TestRTXLidar(omni.kit.test.AsyncTestCase):
+class TestROS2RTXSensor(omni.kit.test.AsyncTestCase):
     # Before running each test
     async def setUp(self):
         await omni.usd.get_context().new_stage_async()
@@ -65,11 +66,9 @@ class TestRTXLidar(omni.kit.test.AsyncTestCase):
         pass
 
     async def test_rtx_lidar_point_cloud(self):
-        # TODO: RTX sensors are not supported on windows yet
-        if sys.platform == "win32":
-            return
         stage = omni.usd.get_context().get_stage()
-        viewport_api = omni.kit.viewport.utility.get_active_viewport()
+
+        viewport_api = get_active_viewport()
         # in order for the sensor to generate data properly we let the viewport know that it should create a buffer for the associated render variable.
         add_aov_to_viewport(viewport_api, "RtxSensorCpu")
 
@@ -79,18 +78,39 @@ class TestRTXLidar(omni.kit.test.AsyncTestCase):
         cube_prim = add_cube(stage, "/World/cube_4", (20, 1, 1), (0, -5, 0), physics=False)
 
         await omni.syntheticdata.sensors.next_sensor_data_async(viewport_api)
-        rv = "RtxSensorCpu"
+        rv = "RtxLidar"
         sensors.get_synthetic_data().activate_node_template(
-            rv + "IsaacReadRTXLidarPointCloud", 0, [viewport_api.get_render_product_path()]
+            rv + "ROS2PublishPointCloud", 0, [viewport_api.get_render_product_path()]
         )
 
         await omni.syntheticdata.sensors.next_sensor_data_async(viewport_api)
 
         _, (_, sensor) = omni.kit.commands.execute("IsaacSensorCreateRtxLidar", path="/sensor", parent=None)
-
         await omni.kit.app.get_app().next_update_async()
         viewport_api.set_active_camera(sensor.GetPath().pathString)
+
+    async def test_rtx_radar_point_cloud(self):
+        stage = omni.usd.get_context().get_stage()
+
+        viewport_api = get_active_viewport()
+        # in order for the sensor to generate data properly we let the viewport know that it should create a buffer for the associated render variable.
+        add_aov_to_viewport(viewport_api, "RtxSensorCpu")
+
+        cube_prim = add_cube(stage, "/World/cube_1", (1, 20, 1), (5, 0, 0), physics=False)
+        cube_prim = add_cube(stage, "/World/cube_2", (1, 20, 1), (-5, 0, 0), physics=False)
+        cube_prim = add_cube(stage, "/World/cube_3", (20, 1, 1), (0, 5, 0), physics=False)
+        cube_prim = add_cube(stage, "/World/cube_4", (20, 1, 1), (0, -5, 0), physics=False)
+
+        await omni.syntheticdata.sensors.next_sensor_data_async(viewport_api)
+        rv = "RtxRadar"
+        sensors.get_synthetic_data().activate_node_template(
+            rv + "ROS2PublishPointCloud", 0, [viewport_api.get_render_product_path()]
+        )
+
+        await omni.syntheticdata.sensors.next_sensor_data_async(viewport_api)
+
+        _, (_, sensor) = omni.kit.commands.execute("IsaacSensorCreateRtxRadar", path="/sensor", parent=None)
         await omni.kit.app.get_app().next_update_async()
-        await omni.kit.app.get_app().next_update_async()
+        viewport_api.set_active_camera(sensor.GetPath().pathString)
 
     pass